_archived/ai-services/requirements.txt
# AI Services Dependencies
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.3
torch==2.1.0
transformers==4.36.0
requests==2.31.0
aiohttp==3.9.0
python-multipart==0.0.6

# Machine Learning
onnx==1.15.0
onnxruntime==1.16.0
joblib==1.3.2

# Cryptography for PQC Signing
cryptography==41.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# NLP
spacy==3.7.2
nltk==3.8.1

# Cryptography and Security
cryptography==41.0.7
pyjwt==2.8.0

# Database and Storage
sqlalchemy==2.0.23
alembic==1.13.0
redis==5.0.1

# Logging and Monitoring
prometheus-client==0.19.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# PDF report generation
fpdf2==2.7.8
mypy==1.7.0

# Reporting
matplotlib==3.7.2
reportlab==4.0.4
